+++ NEWS.md
@@ -1,3 +1,9 @@
+# mvMonitoring Package v 0.2.1
+Last updated: 2023-05-12
+
+This package build is quite stable now. This most recent patch update is due to breaking changes in the `base::density.default()` function as described in <https://bugs.r-project.org/show_bug.cgi?id=18337>.
+
+
 # mvMonitoring Package v 0.0.0.9000
 Last updated: 2017-10-18